Add files via upload
I recently added a CSS stylesheet to my website to make it look better. I also fixed the line breaks in the first paragraph of my index file so that it reads more smoothly. On top of that, I included an image to give the page some visual appeal. These changes are helping to make my website more user-friendly and visually interesting.